+++ src/insert/comments-expanded-advert.ts
@@ -3,6 +3,10 @@ import { adSizes } from 'core/ad-sizes';
 import { AD_LABEL_HEIGHT } from 'core/constants/ad-label-height';
 import { createAdSlot } from 'core/create-ad-slot';
 import { commercialFeatures } from 'lib/commercial-features';
+import { getBreakpoint } from 'lib/detect/detect-breakpoint';
+import { getViewport } from 'lib/detect/detect-viewport';
+import { dfpEnv } from 'lib/dfp/dfp-env';
+import { getAdvertById } from 'lib/dfp/get-advert-by-id';
 import fastdom from '../utils/fastdom-promise';
 import { fillDynamicAdSlot } from './fill-dynamic-advert-slot';
 
@@ -10,7 +14,6 @@ const tallestCommentAd = adSizes.mpu.height + AD_LABEL_HEIGHT;
 const tallestCommentsExpandedAd = adSizes.halfPage.height + AD_LABEL_HEIGHT;
 
 const insertAd = (anchor: HTMLElement) => {
-	log('commercial', 'Inserting comments-expanded advert');
 	const slot = createAdSlot('comments-expanded', {
 		classes: 'comments-expanded',
 	});
@@ -25,9 +28,36 @@ const insertAd = (anchor: HTMLElement) => {
 	stickyContainer.style.flexGrow = '1';
 	stickyContainer.appendChild(adSlotContainer);
 
+	log('commercial', 'Inserting comments-expanded advert');
+
 	return fastdom
 		.mutate(() => {
 			anchor.appendChild(adSlotContainer);
+		})
+		.then(() => fillDynamicAdSlot(slot, false));
+};
+
+const insertAdMobile = (anchor: HTMLElement, id: number) => {
+	const slot = createAdSlot('comments-expanded', {
+		name: `comments-expanded-${id}`,
+		classes: 'comments-expanded',
+	});
+	slot.style.minHeight = `${adSizes.mpu.height + AD_LABEL_HEIGHT}px`;
+
+	const adSlotContainer = document.createElement('div');
+	adSlotContainer.className = 'ad-slot-container';
+	adSlotContainer.style.width = '300px';
+	adSlotContainer.style.margin = '20px auto';
+	adSlotContainer.appendChild(slot);
+
+	const listElement = document.createElement('li');
+	listElement.appendChild(adSlotContainer);
+
+	log('commercial', `Inserting mobile comments-expanded-${id} advert`);
+
+	return fastdom
+		.mutate(() => {
+			anchor.after(listElement);
 		})
 		.then(() => fillDynamicAdSlot(slot, false));
 };
@@ -43,9 +73,20 @@ const getRightColumn = (): HTMLElement => {
 	return rightColumn;
 };
 
+const getCommentsColumn = async (): Promise<HTMLElement> => {
+	return fastdom.measure(() => {
+		const commentsColumn: HTMLElement | null = document.querySelector(
+			'[data-commercial-id="comments-column"]',
+		);
+		if (!commentsColumn) throw new Error('Comments are not expanded.');
+
+		return commentsColumn;
+	});
+};
+
 const isEnoughSpaceForAd = (rightColumnNode: HTMLElement): boolean => {
-	// Only insert a second advert into the right-hand rail if there is enough space.
-	// There is enough space if the right-hand rail is larger than:
+	// Only insert a second advert into the right-hand column if there is enough space.
+	// There is enough space if the right-hand column is larger than:
 	// (the largest possible heights of both adverts) + (the gap between the two adverts)
 	const minHeightToPlaceAd =
 		tallestCommentAd + tallestCommentsExpandedAd + window.innerHeight;
@@ -53,49 +94,103 @@ const isEnoughSpaceForAd = (rightColumnNode: HTMLElement): boolean => {
 	return rightColumnNode.offsetHeight >= minHeightToPlaceAd;
 };
 
-const createResizeObserver = (rightColumnNode: HTMLElement) => {
-	// When the comments load and are rendered, the height of the right column
-	// will expand and there might be enough space to insert the ad.
-	const resizeObserver = new ResizeObserver(() => {
-		if (isEnoughSpaceForAd(rightColumnNode)) {
-			void insertAd(rightColumnNode);
+const isEnoughCommentsForAd = (commentsColumn: HTMLElement): boolean =>
+	commentsColumn.childElementCount >= 5;
 
-			resizeObserver.unobserve(rightColumnNode);
-		}
-	});
+const commentsExpandedAdsAlreadyExist = (): boolean => {
+	const commentsExpandedAds = document.querySelectorAll(
+		'.ad-slot--comments-expanded',
+	);
 
-	resizeObserver.observe(rightColumnNode);
+	return commentsExpandedAds.length > 0 ? true : false;
 };
 
-/**
- * Create a comments-expanded ad immediately if there is enough space for it. If not, then it
- * is possible that we are still waiting for the Discussion API to load the comments, so we
- * wait for the comments to load before checking again whether there is enough space to load an ad.
- */
-const handleCommentsExpandedEvent = (): void => {
+const removeMobileCommentsExpandedAds = (): Promise<void> => {
+	const currentBreakpoint = getBreakpoint(getViewport().width);
+	if (currentBreakpoint !== 'mobile') {
+		return Promise.resolve();
+	}
+
+	const commentsExpandedAds = document.querySelectorAll(
+		'.ad-slot--comments-expanded',
+	);
+
+	return fastdom.mutate(() =>
+		commentsExpandedAds.forEach((node) => {
+			log('commercial', `Removing ad slot: ${node.id}`);
+			const advert = getAdvertById(node.id);
+			if (advert) {
+				window.googletag.destroySlots([advert.slot]);
+			}
+			node.remove();
+			dfpEnv.adverts.delete(node.id);
+			dfpEnv.advertsToLoad = dfpEnv.advertsToLoad.filter(
+				(_) => _ !== advert,
+			);
+		}),
+	);
+};
+
+const handleCommentsLoadedEvent = (): void => {
+	const rightColumnNode = getRightColumn();
+
+	if (isEnoughSpaceForAd(rightColumnNode)) {
+		void insertAd(rightColumnNode);
+	}
+};
+
+const handleCommentsLoadedMobileEvent = async (): Promise<void> => {
+	const commentsColumn = await getCommentsColumn();
+
+	// On frontend-rendered pages, there is a merchandising-high ad below the comments ad.
+	// We want a sufficient amount of content between these two ads.
+	const isDcr = window.guardian.config.isDotcomRendering;
+	const minCommentsBelowAd = isDcr ? 1 : 3;
+
+	if (
+		isEnoughCommentsForAd(commentsColumn) &&
+		!commentsExpandedAdsAlreadyExist()
+	) {
+		let counter = 0;
+		for (let i = 0; i < commentsColumn.childElementCount; i++) {
+			if (
+				commentsColumn.children[i] &&
+				(i - 3) % 5 === 0 && // The fourth comment and then every fifth comment
+				i + minCommentsBelowAd < commentsColumn.childElementCount
+			) {
+				counter++;
+				const childElement = commentsColumn.children[i] as HTMLElement;
+				void insertAdMobile(childElement, counter);
+			}
+		}
+	}
+};
+
+export const initCommentsExpandedAdverts = (): Promise<void> => {
 	if (!commercialFeatures.commentAdverts) {
 		log(
 			'commercial',
 			'Adverts in comments are disabled in commercialFeatures',
 		);
-		return;
+		return Promise.resolve();
 	}
 
-	const rightColumnNode = getRightColumn();
+	document.addEventListener('comments-loaded', () => {
+		const currentBreakpoint = getBreakpoint(getViewport().width);
+		if (currentBreakpoint === 'mobile') {
+			void handleCommentsLoadedMobileEvent();
+		} else {
+			void handleCommentsLoadedEvent();
+		}
+	});
 
-	if (isEnoughSpaceForAd(rightColumnNode)) {
-		void insertAd(rightColumnNode);
-		return;
-	}
-
-	// Watch the right column and try to insert an ad when the comments are loaded.
-	createResizeObserver(rightColumnNode);
-};
-
-export const initCommentsExpandedAdverts = (): Promise<void> => {
-	document.addEventListener('comments-expanded', () =>
-		handleCommentsExpandedEvent(),
-	);
+	/**
+	 * If the page of comments is changed, or the ordering is updated, etc,
+	 * we need to remove the existing slots and create new slots.
+	 */
+	document.addEventListener('comments-state-change', () => {
+		void removeMobileCommentsExpandedAds();
+	});
 
 	return Promise.resolve();
 };