+++ src/lib/header-bidding/prebid-types.ts
@@ -5,7 +5,9 @@ export type HeaderBiddingSize = AdSize;
 export type HeaderBiddingSlotName =
 	| 'banner'
 	| 'comments'
+	| 'comments-expanded'
 	| 'crossword-banner'
+	| 'crossword-banner-mobile'
 	| 'mobile-sticky'
 	| 'mostpop'
 	| 'right'